testsuite: create separate c files, adapt build system
