+++ makefile.msvc
@@ -67,8 +67,8 @@ $(LIBMAIN_S): $(OBJECTS)
 	lib /out:$(LIBMAIN_S) $(OBJECTS)
 
 #Build test_standalone suite
-test.exe: $(LIBMAIN_S) demo/demo.c
-	cl $(LTM_CFLAGS) $(TOBJECTS) $(LIBMAIN_S) $(LTM_LDFLAGS) demo/demo.c /DLTM_DEMO_TEST_VS_MTEST=0 /Fe$@
+test.exe: $(LIBMAIN_S) demo/main.c demo/opponent.c demo/test.c
+	cl $(LTM_CFLAGS) $(TOBJECTS) $(LIBMAIN_S) $(LTM_LDFLAGS) demo/main.c demo/opponent.c demo/test.c /DLTM_DEMO_TEST_VS_MTEST=0 /Fe$@
 	@echo NOTICE: start the tests by launching test.exe
 
 test_standalone: test.exe
