+++ makefile.unix
@@ -79,8 +79,8 @@ $(LIBMAIN_S): $(OBJECTS)
 	$(RANLIB) $@
 
 #Build test_standalone suite
-test: $(LIBMAIN_S) demo/demo.c
-	$(CC) $(LTM_CFLAGS) $(LTM_LDFLAGS) demo/demo.c $(LIBMAIN_S) -DLTM_DEMO_TEST_VS_MTEST=0 -o $@
+test: demo/main.c demo/opponent.c demo/test.c $(LIBMAIN_S)
+	$(CC) $(LTM_CFLAGS) $(LTM_LDFLAGS) $^ -DLTM_DEMO_TEST_VS_MTEST=0 -o $@
 	@echo "NOTICE: start the tests by: ./test"
 
 test_standalone: test
@@ -90,7 +90,7 @@ all: $(LIBMAIN_S) test_standalone
 #NOTE: this makefile works also on cygwin, thus we need to delete *.exe
 clean:
 	-@rm -f $(OBJECTS) $(LIBMAIN_S)
-	-@rm -f demo/demo.o test test.exe
+	-@rm -f demo/main.o demo/opponent.o demo/test.o test test.exe
 
 #Install the library + headers
 install: $(LIBMAIN_S)
